requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.4.2
cycler==0.10.0
django==2.1.3
idna==2.8
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.16.0
opencv-python==4.0.0.21
pillow==5.4.1
pycparser==2.19
pymysql==0.9.2
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sklearn==0.0
